% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-age.R
\docType{data}
\name{age}
\alias{age}
\title{Mice Age}
\format{
A data frame with 29 rows and 134 columns.
}
\source{
\emph{Metabolomics Workbench}
https://www.metabolomicsworkbench.org/data/showfile_t.php?RA=174.45.40.196&DF=MSdata_ST001703_1.txt
}
\usage{
age
}
\description{
Log-transformed and standardized named metabolite concentrations of brain tissue samples
from 29 mice of different ages. Groups are "young": < 15 months,
"mid age": > = 15 months = < 20 months, "old": > 20 months
}
\references{
This data is available at the NIH Common Fund's National Metabolomics
Data Repository (NMDR) website, the Metabolomics Workbench,
https://www.metabolomicsworkbench.org, where it has been assigned Project ID PR001089.
The data can be accessed directly via it's Project DOI: 10.21228/M8XT42.
This work is supported by NIH grant, U2C- DK119886.
}
\keyword{datasets}
