% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diet.R
\docType{data}
\name{diet}
\alias{diet}
\title{Cow diet}
\format{
A data frame with 39 rows and 48 columns.
}
\source{
\emph{Metaboanalyst 5.0}
https://www.xialab.ca/api/download/metaboanalyst/cow_diet.csv
}
\usage{
diet
}
\description{
Log-transformed and standardized named metabolite concentrations of 39 rumen
samples measured by proton NMR for dairy cows fed with different proportions of
barley grain. Group labels 0, 15, 30, 45
}
\references{
Ametaj, B.N., Zebeli, Q., Saleem, F. et al. Metabolomics reveals
unhealthy alterations in rumen metabolism with increased proportion of cereal
grain in the diet of dairy cows. Metabolomics 6, 583–594 (2010).
https://doi.org/10.1007/s11306-010-0227-6
}
\keyword{datasets}
