% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiermt.R
\name{hiermt}
\alias{hiermt}
\title{Title: Hierarchical multiple testing}
\usage{
hiermt(formula, data, global_test, alpha)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame in containing the variables named in \code{formula}.}

\item{global_test}{Global test to use when testing the intersection of hypotheses.}

\item{alpha}{Probability of Type I error.}
}
\description{
Description: Calculate and visualize hierarchically bonferroni-adjusted
p-values.
}
\examples{
set.seed(1)
library(MASS)
df <- data.frame(mvrnorm(n=80,mu=rep(0,100),Sigma=diag(100)), label = rep(c(0,1),40))
hiermt(formula=.~label, data = df, global_test="bonferroni", alpha = 0.05)
}
