% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hiermt.R
\name{plot.hiermt}
\alias{plot.hiermt}
\title{Plot Hierarchical Model Testing Results}
\usage{
\method{plot}{hiermt}(x, ...)
}
\arguments{
\item{x}{An object of class 'hiermt'.}

\item{...}{Additional arguments.}
}
\description{
This function plots the results of hierarchical model testing.
}
\examples{
set.seed(1)
n <- 50
df <- data.frame(y1 = rnorm(n), y2 = rnorm(n), x = sample(rep(0:1, c(n / 2, n / 2))))
hmt <- hiermt(formula = cbind(y1, y2) ~ x, data = df, global_test = "bonferroni", alpha = 0.05)
plot(hmt)
}
