% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.hiermt.R
\name{summary.hiermt}
\alias{summary.hiermt}
\title{Summary of Hierarchical Model Testing}
\usage{
\method{summary}{hiermt}(object, ...)
}
\arguments{
\item{object}{An object of class 'hiermt'.}

\item{...}{Additional arguments.}
}
\value{
\code{summary} summary of results
}
\description{
This function produces the summary of hierarchical model testing.
}
\examples{
set.seed(1)
n <- 50
df <- data.frame(y1 = rnorm(n), y2 = rnorm(n), x = sample(rep(0:1, c(n / 2, n / 2))))
hmt <- hiermt(formula = cbind(y1, y2) ~ x, data = df, global_test = "bonferroni", alpha = 0.05)
plot(hmt)

}
