% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiermt.R
\name{hiermt}
\alias{hiermt}
\title{Title: Hierarchical multiple testing}
\usage{
hiermt(pvalues, test_stats, cor_mat, global_test, alpha = 1, method)
}
\arguments{
\item{pvalues}{A numeric vector of p-values of length Q.}

\item{test_stats}{A numeric vector of test statistics of length Q.}

\item{cor_mat}{A Q*Q square matrix of the correlations between the p-values or test statistics.}

\item{global_test}{Global test to use when testing the intersection of hypotheses. This should be one of "bonferonni", "ghc", "gbj".}

\item{alpha}{Probability of Type I error. Default is set to 1.}

\item{method}{Multiple testing adjustment method, a character string. "Meinshausen's method", "Inheritance procedure"}
}
\description{
Description: Calculate and visualize correlation-based hierarchically adjusted p-values.
}
\examples{
set.seed(1)
Q <- 50
test_stats <- rnorm(50, mean = c(rep(0, 25), rep(3, 25)))
pvalues <- 2*pnorm(sort(-abs(test_stats)))
cor_mat <- matrix(data= 0.3, nrow=Q, ncol=Q)
diag(cor_mat) <- 1
hiermt(pvalues, test_stats, cor_mat, global_test="bonferroni", alpha = 0.05, method=NULL)
}
