% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiermt.R
\name{hiermt}
\alias{hiermt}
\title{Hierarchical multiple testing}
\usage{
hiermt(pvalues, test_stats, cor_mat, global_test = "ghc", alpha = 1)
}
\arguments{
\item{pvalues}{A numeric vector of p-values of length Q.}

\item{test_stats}{A numeric vector of test statistics of length Q.}

\item{cor_mat}{A Q*Q square matrix of the correlations between the p-values or test statistics.}

\item{global_test}{Global test to use when testing the intersection of hypotheses. This should be one of "bonferonni", "ghc", "gbj".}

\item{alpha}{Probability of Type I error. Default is set to 1.}
}
\value{
\code{dend} A dendrogram with the hierarchically-adjusted p-values.

\code{leaves} A dataframe containing the adjusted p-values for the individual hypotheses.

\code{all} A dataframe containing all attributes of each node in the hierarchy.
}
\description{
Calculate and visualize correlation-based hierarchically adjusted p-values.
}
\examples{
set.seed(1)
Q <- 50
test_stats <- rnorm(Q, mean = c(rep(0, Q/2), rep(3, Q/2)))
pvalues <- 2*pnorm(sort(-abs(test_stats)))
cor_mat <- clusterGeneration::rcorrmatrix(Q)
hiermt(pvalues, test_stats, cor_mat, global_test="bonferroni", alpha = 0.05)
}
