% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiermt.R
\name{hiermt}
\alias{hiermt}
\title{Fit linear models and hierarchically adjust p-values}
\usage{
hiermt(
  formula,
  data = NULL,
  global_test = "ghc",
  alpha = 0.05,
  linkage = "ward.D2",
  mult_comp = FALSE
)
}
\arguments{
\item{formula}{A model formula object. Left hand side contains response variable(s) and right hand side contains explanatory variable(s).}

\item{data}{A data frame containing variables listed in the formula.}

\item{global_test}{Global test to use when testing the hypotheses. This should be one of "bonferroni", "ghc", "gbj".}

\item{alpha}{Probability of Type I error. Default is set to 1.}

\item{linkage}{Agglomeration method used for hierachical clustering in hclust. Defaults to Ward's method.}

\item{mult_comp}{Logical: TRUE or FALSE. Whether to compute multiple comparisons.}
}
\value{
\code{dend} A dendrogram with the hierarchically-adjusted p-values.

\code{leaves} A dataframe containing the adjusted p-values for the terminal node hypotheses.

\code{all} A dataframe containing all attributes of each node in the hierarchy.
}
\description{
hiermt is used to fit multiple univariate linear models and perform hierarchical-based multiple testing adjustment to the resulting p-values.
}
\examples{
set.seed(1)
n <- 50
df <- data.frame(y1 = rnorm(n), y2 = rnorm(n), x = sample(rep(0:1, c(n / 2, n / 2))))
hiermt(formula = cbind(y1, y2) ~ x, data = df, global_test = "bonferroni", alpha = 0.05)
hiermt(formula = . ~ x, data = df, global_test = "bonferroni", alpha = 0.05)
}
